refactor: extract the function of FactCreationModal to FactCreationBox to reused it in FactImportModal
