+++ package.json
@@ -18,6 +18,7 @@
     "@tanstack/react-query": "^5.71.0",
     "@types/lodash": "^4.17.16",
     "lodash": "^4.17.21",
+    "motion": "^12.9.1",
     "next": "15.2.4",
     "posthog-js": "^1.234.1",
     "react": "19.1.0",
